hms/hms_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8

gunicorn==20.1.0
dj-database-url==1.0.0
whitenoise==6.5.0
django-storages[boto3]==1.13.1
boto3==1.26.0
